Rest Template for Users-UserActivities initial

+++ users/src/main/java/com/service/controllers/UserController.java
@@ -1,5 +1,14 @@
 package com.service.controllers;
 
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.MediaType;
 
 import com.service.customExceptions.UserNotFoundException;
 import com.service.entities.User;
@@ -9,6 +18,10 @@ import com.service.repositories.UserRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.UriComponentsBuilder;
+
+import javax.validation.Valid;
 
 
 @RestController
@@ -16,34 +29,82 @@ public class UserController {
 
     private final UserRepository userRepository;
 
+    @Autowired
+    RestTemplate restTemplate;
 
-    UserController(UserRepository repository, UserDetailRepository detailsRepository) {
+
+    UserController(UserRepository repository) {
         this.userRepository = repository;
     }
 
-    @GetMapping("/users")
-    public Iterable<User> getAllUsers() {
-        return userRepository.findAll();
+    // Get all users from the database
+    @RequestMapping(method = RequestMethod.GET, value = "/users")
+    public ResponseEntity<List<User>> getAllUsers() {
+        List<User> users = (List<User>) this.userRepository.findAll();
+        if (users.isEmpty()) {
+            // check if there are no items in the list
+            return new ResponseEntity(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<List<User>>(users, HttpStatus.OK);
     }
 
-    @PostMapping("/addUser")
-    User addNewUser(@RequestBody User newUser) {
-
-        return userRepository.save(newUser);
+    // Get only one user from the database, specified by ID
+    @RequestMapping(method = RequestMethod.GET, value = "/{userId}")
+    ResponseEntity<?> getUserById(@PathVariable Long userid) {
+        Optional<User> user = this.userRepository.findById(userid);
+        if (!user.isPresent()) {
+            // may implement custom exception, better practice
+            return new ResponseEntity(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<Optional<User>>(user, HttpStatus.OK);
     }
 
-    @GetMapping("/user/{id}")
-    User getUserById(@PathVariable Long id) {
+    // Add new user to the database
+    @RequestMapping(value = "/add", method = RequestMethod.POST)
+    public ResponseEntity<?> createLocation(@Valid @RequestBody User user, UriComponentsBuilder builder) {
 
-        return userRepository.findById(id)
-                .orElseThrow(() -> new UserNotFoundException(id));
+        // Check if username already exists
+        List<User> users = userRepository.findAll();
+        if (!users.isEmpty()) {
+            for (User u : users) {
+                if (u.getUsername() == user.getUsername())
+                    return new ResponseEntity("User with this username already exists!", HttpStatus.CONFLICT);
+            }
+        }
+
+        // Need to check if update to UserDetails TABLE is required...
+        return new ResponseEntity<String>(HttpStatus.CREATED);
     }
 
-    @DeleteMapping("/removeUser/{id}")
-    ResponseEntity<String> deleteUserById(@PathVariable Long id) {
+    //Update existing user
+    @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
+    public ResponseEntity<?> updateUser(@PathVariable("id") long id,@Valid @RequestBody User newUser) {
+
+        Optional<User> existingUser = userRepository.findById(id);
+
+        if (!existingUser.isPresent()) {
+            return new ResponseEntity("Unable to upate. Country with id " + id + " not found.",
+                    HttpStatus.NOT_FOUND);
+        }
+
+        existingUser.get().setUsername(newUser.getUsername());
+
+        userRepository.save(existingUser.get());
+        return new ResponseEntity<Optional<User>>(existingUser, HttpStatus.OK);
+    }
+
+    // Delete user from the database
+    @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
+    public ResponseEntity<?> deleteUser(@PathVariable("id") Long id) {
+        // Validate that user exists in the Db
+        Optional<User> user = userRepository.findById(id);
+        if (!user.isPresent())
+            return new ResponseEntity("User with specific ID: " + id + " was NOT found!", HttpStatus.NOT_FOUND);
         userRepository.deleteById(id);
-        return new ResponseEntity<>("User with id: " + id + " was deleted!", HttpStatus.OK);
-
+        return new ResponseEntity<User>(HttpStatus.NO_CONTENT);
     }
+
+    // May implement DROP ALL USERS ?!
+
 
 }
